+++ usecases/blea-gov-base-ct/parameter.ts
@@ -4,6 +4,8 @@ export interface AppParameter {
   env?: Environment;
   envName: string;
   securityNotifyEmail: string;
+  securitySlackWorkspaceId?: string; // required if deploy via CLI
+  securitySlackChannelId?: string; // required if deploy via CLI
 }
 
 export interface PipelineParameter {
@@ -21,6 +23,8 @@ export interface PipelineParameter {
 export const devParameter: AppParameter = {
   envName: 'Development',
   securityNotifyEmail: 'notify-security@example.com',
+  securitySlackWorkspaceId: 'T8XXXXXXX',
+  securitySlackChannelId: 'C00XXXXXXXX',
   // env: { account: '123456789012', region: 'ap-northeast-1' },
 };